Fitness.cpp completed

+++ src/fitness.cpp
@@ -1,6 +1,6 @@
 #include "fitness.h"
 
-float fitness(vec_1D &solution)
+int fitness(vec_1D &solution)
 {
     int fitnessValue=0;
     fitnessValue = pow((solution[0]-2),2) + pow(solution[1]-3,4)\
